refactor : 패키징처리
