+++ src/main/kotlin/blackjack/domain/Score.kt
@@ -1,5 +1,8 @@
 package blackjack.domain
 
+import blackjack.domain.card.ACE
+import blackjack.domain.card.Cards
+
 /**
  * Created by Jaesungchi on 2022.06.14..
  */
